✨ Settings menu with ignorables, synced with right-click-ignoring
The ignored rules are stored in the LO/OO user registry, loaded on
startup, and respected when grammar-checking.

Settings are shown as two panes with buttons between (ignored vs
included errors). LO/OO doesn't have an API for checkbox lists,
meaning we'd have to
[make one from scratch](https://stackoverflow.com/questions/49732534/creating-a-checklistbox-in-libreoffice-from-python)
which seems like waste of time, while the two-pane alternative turned
out pretty usable :)

We also support incomplete errors.xml's: If a user right-clicks to
ignore an error, we include it in settings, just showing the id, since
there's no message.

The ignorables are unique'd using their error messages, so if several
errors have the same message, they are treated the same
wrt. ignoring. Currently, some different errors have identical
messages in sme – if we want users to differentiate them, we need to
provide unique messages in errors.xml

Also added some mypy type hints (though not sure if we can do that for
the UNO stuff?).